- cross-platform docs for `node-archive` trait - replace ad-hoc conditional compilation with structured `cfg_if!` macro - start replacing `not(windows)` config logic with `unix`
