+++ crates/notion-core/Cargo.toml
@@ -4,7 +4,7 @@ version = "0.1.0"
 authors = ["David Herman <david.herman@gmail.com>"]
 
 [features]
-universal-docs = []
+universal-docs = ["node-archive/universal-docs"]
 
 [dependencies]
 toml = "0.4"
@@ -24,7 +24,4 @@ semver = "0.9.0"
 cmdline_words_parser = "0.0.2"
 reqwest = "0.7.3"
 cfg-if = "0.1"
-
-[target.'cfg(windows)'.dependencies]
-zip = "0.2.6"
 winfolder = "0.1"
